Move into section folder and create project section
