+++ src/components/section/projects.js
@@ -1,0 +1,7 @@
+import React from "react"
+
+const Projects = () => {
+  return <div>Projects</div>
+}
+
+export default Projects
